DebtController has been updated.

+++ src/main/java/com/furkanyesilyurt/fourthHomework/dto/debt/DebtDTO.java
@@ -15,6 +15,6 @@ public class DebtDTO {
     private Date expiryDate;
     private DebtType debtType;
     private Long debt;
-    private Long delayDebt;
+    private Double delayDebt;
 
 }
