+++ src/main/java/com/furkanyesilyurt/fourthHomework/dto/debt/DebtExpiryDateDTO.java
@@ -16,6 +16,6 @@ public class DebtExpiryDateDTO {
     private Double remainingDebt;
     private Date expiryDate;
     private Long debt;
-    private Long delayDebt;
+    private Double delayDebt;
 
 }
